add component to calculate feature importance
