+++ responsibleai/model_monitor/components/feature_importance/asset.yaml
@@ -1,0 +1,2 @@
+type: component
+spec: spec.yaml
