+++ responsibleai/model_monitor/components/feature_importance/spec.yaml
@@ -1,22 +1,19 @@
 $schema: http://azureml/sdk-2-0/CommandComponent.json
-name: feature_attribution_drift
+name: feature_importance
 version: 0.0.1
-display_name: feature_attribution_drift
+display_name: feature_importance
 is_deterministic: False
 type: command
-description: Feature attribution drift using model monitoring.
+description: Feature importance using model monitoring.
 inputs:
   baseline_data:
     type: mltable
-  production_data:
-    type: mltable
 outputs:
-  normalized_discounted_cumulative_gain:
+  feature_importance:
     type: mltable
 code: ./
 environment: azureml://registries/azureml/environments/responsibleai-ubuntu20.04-py38-cpu/versions/10
 resources:
   instance_count: 1
-command: python compute_attribution_drift.py
+command: python compute_feature_importance.py
   --baseline_data ${{inputs.baseline_data}}
-  --production_data ${{inputs.production_data}}
